+++ src/entities/contents/main/api/index.ts
@@ -5,5 +5,4 @@ export { fetchDetailData } from './fetchDetailData';
 export { fetchSections } from './fetchSections';
 export { fetchOttPlatforms } from './fetchOttPlatforms';
 export { fetchContents } from './fetchContents';
-export { fetchSearchContents } from '../../../../shared/api/search/fetchSearchContents';
 export { fetchOrder } from './fetchOrder';
